move syevd +flip functions in misc and other cov based methods updates
